refactor: dedupe image binarization + simplify resize.mjs
- New src/image-processing/binarize.mjs: computeOtsuThreshold,
  applyThreshold, sharpen, padAndScale, shouldInvert -- were
  byte-identical copies in resize.mjs and ocr-preprocessing.mjs (or, for
  padAndScale/shouldInvert, one parameter away from identical). One
  implementation now.
- Along the way: fixed resize.mjs's padAndScale call site passing 4 args
  to a 3-param function (the 4th, scaleFactor, was silently dropped --
  harmless since it read the same value off the closure-captured
  preprocessConfig anyway, but a real signature-mismatch smell).
- resize.mjs's GetAlteredDimensions: collapsed 5 near-identical
  if-type-then-return blocks into one lookup against the constants
  object (name/type/art/flavor/setSymbol), with left/top defaulting to
  constants.borderPercent unless the type overrides it.
- resize.mjs's binaryDilate/binaryErode: same 3x3-neighborhood sweep
  differing only in max-vs-min -- merged into one morphologicalOp(img,
  iterations, combine, initial) helper.

Verified: pnpm check (289 tests) and pnpm test:regression (71/71
blocking, unchanged from baseline) both green -- behavior-preserving.
